+++ src/components/radio-field/radio-field.define.js
@@ -1,0 +1,3 @@
+import { define } from './radio-field.js';
+
+define();